+++ language/en/info_acp_imageupload.php
@@ -14,7 +14,7 @@ if (!defined('IN_PHPBB'))
 
 if (empty($lang) || !is_array($lang))
 {
-	$lang = array();
+	$lang = [];
 }
 
 // DEVELOPERS PLEASE NOTE
@@ -33,7 +33,7 @@ if (empty($lang) || !is_array($lang))
 // ’ » “ ” …
 //
 
-$lang = array_merge($lang, array(
+$lang = array_merge($lang, [
 	'ACP_IMAGE_UPLOAD'			=> 'Image Upload',
 	'ACP_IMAGE_UPLOAD_CONFIG'	=> 'Configuration',
 	// Log
@@ -41,4 +41,4 @@ $lang = array_merge($lang, array(
 	'LOG_IMAGEUPLOAD_SETTINGS'	=> '<strong>Image Upload configuration changed</strong>',
 	'LOG_IMAGEUPLOAD_DELETED'	=> '<strong>Deleted image</strong><br>» %1$s',
 	'LOG_USER_IMAGE_DELETED'	=> '<strong>Image %s deleted with real name %s on server</strong>',
-));
+]);